add search functionality in report management and prevent scrolling load while searching
Added a search functionality in the 'manage-work' panel, allowing reports to be filtered based on a search query. This was achieved by adding a search form which interacts with the new 'searchReports' method in the data service. To prevent conflicts with the existing scrolling load feature, loadMore was updated to only fire if 'scrolling' is true, which is set to false during a search. Miscellaneous style adjustments and class name changes were also implemented.