update option label in dropdown menu
Changed the displayed value for the selected option in the dropdown menu in the edit_report component. Instead of just showing the 'machine_name', now it shows 'machine_code', 'machine_brand', and 'machine_name'. This change was made to add more context and give clear identification of the chosen machine.